Hack org-refile-targets in org-goto-from-anywhere.
bkuhn/org-goto-from-anywhere now takes a universal argument, which will
hack the :level setting of org-refile-targets, and thus allowing the
completion list when org-refile-get-location is called to include
different sets of items based on levels.

I noticed an interesting fact in implementing this: care must be taken
to clear out org-refile-target-table again if you've made locally scoped
changes to org-refile-targets, because it seems there is some caching
going on (*unrelated* to the org-refile-cache stuff).

However, I only do that "clearing" when I really did muck with
org-refile-targets locally.