Use original-buffer so full context available.
org-capture, by the time %(sexp) in the template are called, has
switched the buffer to the capture buffer.

So, since this function seeks information from the IRC log buffer that
may not yet be present in the current buffer, switch back to the
original-buffer, assuming that property is set, and do the operations
there.